FIX Better handling of re-sampling saved images
On write of a content service asset, clear out all resamples and mark for
regeneration. Ensures that a 'save' of an item will re-generate its underlying
resamples on their next request.

Additionally, rather than immediately deleting a resampled asset from the
remote storage, simply mark it as deleted. If the resampled asset was referred
to by cached content, we don't want that removed until _after_ the new
asset is picked up in that cached content reference

Finally, added a job to cleanup those assets that have been marked for deletion